crates/gosub_html5/tests/data/html5lib-tests/tree-construction/tests19.dat

#data
<!doctype html><table><i>a<b>b<div>c<a>d</i>e</b>f
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,29): foster-parenting-start-tag
(1,30): foster-parenting-character
(1,35): foster-parenting-start-tag
(1,36): foster-parenting-character
(1,39): foster-parenting-start-tag
(1,40): foster-parenting-character
(1,44): foster-parenting-end-tag
(1,44): adoption-agency-1.3
(1,44): adoption-agency-1.3
(1,45): foster-parenting-character
(1,49): foster-parenting-end-tag
(1,49): adoption-agency-1.3
(1,49): adoption-agency-1.3
(1,50): foster-parenting-character
(1,50): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|     <div>
|       <b>
|         <i>
|           "c"
|           <a>
|             "d"
|         <a>
|           "e"
|       <a>
|         "f"
|     <table>

#data
<!doctype html><i>a<b>b<div>c<a>d</i>e</b>f
#errors
(1,37): adoption-agency-1.3
(1,37): adoption-agency-1.3
(1,42): adoption-agency-1.3
(1,42): adoption-agency-1.3
(1,43): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|     <div>
|       <b>
|         <i>
|           "c"
|           <a>
|             "d"
|         <a>
|           "e"
|       <a>
|         "f"

#data
<!doctype html><table><i>a<b>b<div>c</i>
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,29): foster-parenting-start-tag
(1,30): foster-parenting-character
(1,35): foster-parenting-start-tag
(1,36): foster-parenting-character
(1,40): foster-parenting-end-tag
(1,40): adoption-agency-1.3
(1,40): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|       <div>
|         <i>
|           "c"
|     <table>

#data
<!doctype html><table><i>a<div>b<tr>c<b>d</i>e
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,31): foster-parenting-start-tag
(1,32): foster-parenting-character
(1,37): foster-parenting-character
(1,40): foster-parenting-start-tag
(1,41): foster-parenting-character
(1,45): foster-parenting-end-tag
(1,45): adoption-agency-1.3
(1,46): foster-parenting-character
(1,46): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <div>
|         "b"
|     <i>
|       "c"
|       <b>
|         "d"
|     <b>
|       "e"
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><table><td><table><i>a<div>b<b>c</i>d
#errors
(1,26): unexpected-cell-in-table-body
(1,36): foster-parenting-start-tag
(1,37): foster-parenting-character
(1,42): foster-parenting-start-tag
(1,43): foster-parenting-character
(1,46): foster-parenting-start-tag
(1,47): foster-parenting-character
(1,51): foster-parenting-end-tag
(1,51): adoption-agency-1.3
(1,51): adoption-agency-1.3
(1,52): foster-parenting-character
(1,52): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <i>
|               "a"
|             <div>
|               <i>
|                 "b"
|                 <b>
|                   "c"
|               <b>
|                 "d"
|             <table>

#data
<!doctype html><body><bgsound>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <bgsound>

#data
<!doctype html><body><basefont>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <basefont>

#data
<!doctype html><a><b></a><basefont>
#errors
(1,25): adoption-agency-1.3
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <basefont>

#data
<!doctype html><a><b></a><bgsound>
#errors
(1,25): adoption-agency-1.3
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <bgsound>

#data
<!doctype html><figcaption><article></figcaption>a
#errors
(1,49): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <figcaption>
|       <article>
|     "a"

#data
<!doctype html><summary><article></summary>a
#errors
(1,43): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <summary>
|       <article>
|     "a"

#data
<!doctype html><p><a><plaintext>b
#errors
(1,32): unexpected-end-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <a>
|     <plaintext>
|       <a>
|         "b"

#data
<!DOCTYPE html><div>a<a></div>b<p>c</p>d
#errors
(1,30): end-tag-too-early
(1,40): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       "a"
|       <a>
|     <a>
|       "b"
|       <p>
|         "c"
|       "d"
